reorg a bit
