--- apps/apps.jl
@@ -1,8 +1,0 @@
-module FinchApps
-
-using Finch
-
-include("graphs.jl")
-include("linalg.jl")
-
-end
