+++ benchmark/benchmarks.jl
@@ -10,7 +10,11 @@ using BenchmarkTools
 using MatrixDepot
 using SparseArrays
 
-include(joinpath(@__DIR__, "../apps/apps.jl"))
+include(joinpath(@__DIR__, "../docs/examples/bfs.jl"))
+include(joinpath(@__DIR__, "../docs/examples/pagerank.jl"))
+include(joinpath(@__DIR__, "../docs/examples/shortest_paths.jl"))
+include(joinpath(@__DIR__, "../docs/examples/spgemm.jl"))
+include(joinpath(@__DIR__, "../docs/examples/triangle_counting.jl"))
 
 SUITE = BenchmarkGroup()
 
@@ -43,18 +47,18 @@ SUITE["graphs"] = BenchmarkGroup()
 
 SUITE["graphs"]["pagerank"] = BenchmarkGroup()
 for mtx in ["SNAP/soc-Epinions1", "SNAP/soc-LiveJournal1"]
-    SUITE["graphs"]["pagerank"][mtx] = @benchmarkable FinchApps.pagerank($(pattern!(fiber(SparseMatrixCSC(matrixdepot(mtx)))))) 
+    SUITE["graphs"]["pagerank"][mtx] = @benchmarkable pagerank($(pattern!(fiber(SparseMatrixCSC(matrixdepot(mtx)))))) 
 end
 
 SUITE["graphs"]["bfs"] = BenchmarkGroup()
 for mtx in ["SNAP/soc-Epinions1", "SNAP/soc-LiveJournal1"]
-    SUITE["graphs"]["bfs"][mtx] = @benchmarkable FinchApps.bfs($(fiber(SparseMatrixCSC(matrixdepot(mtx))))) 
+    SUITE["graphs"]["bfs"][mtx] = @benchmarkable bfs($(fiber(SparseMatrixCSC(matrixdepot(mtx))))) 
 end
 
 SUITE["graphs"]["bellmanford"] = BenchmarkGroup()
 for mtx in ["Newman/netscience", "SNAP/roadNet-CA"]
     A = redefault!(fiber(SparseMatrixCSC(matrixdepot(mtx))), Inf)
-    SUITE["graphs"]["bellmanford"][mtx] = @benchmarkable FinchApps.bellmanford($A)
+    SUITE["graphs"]["bellmanford"][mtx] = @benchmarkable bellmanford($A)
 end
 
 SUITE["matrices"] = BenchmarkGroup()
@@ -62,19 +66,19 @@ SUITE["matrices"] = BenchmarkGroup()
 SUITE["matrices"]["ATA_spgemm_inner"] = BenchmarkGroup()
 for mtx in []#"SNAP/soc-Epinions1", "SNAP/soc-LiveJournal1"]
     A = fiber(permutedims(SparseMatrixCSC(matrixdepot(mtx))))
-    SUITE["matrices"]["ATA_spgemm_inner"][mtx] = @benchmarkable FinchApps.spgemm_inner($A, $A) 
+    SUITE["matrices"]["ATA_spgemm_inner"][mtx] = @benchmarkable spgemm_inner($A, $A) 
 end
 
 SUITE["matrices"]["ATA_spgemm_gustavson"] = BenchmarkGroup()
 for mtx in ["SNAP/soc-Epinions1"]#], "SNAP/soc-LiveJournal1"]
     A = fiber(SparseMatrixCSC(matrixdepot(mtx)))
-    SUITE["matrices"]["ATA_spgemm_gustavson"][mtx] = @benchmarkable FinchApps.spgemm_gustavson($A, $A) 
+    SUITE["matrices"]["ATA_spgemm_gustavson"][mtx] = @benchmarkable spgemm_gustavson($A, $A) 
 end
 
 SUITE["matrices"]["ATA_spgemm_outer"] = BenchmarkGroup()
 for mtx in ["SNAP/soc-Epinions1"]#, "SNAP/soc-LiveJournal1"]
     A = fiber(SparseMatrixCSC(matrixdepot(mtx)))
-    SUITE["matrices"]["ATA_spgemm_outer"][mtx] = @benchmarkable FinchApps.spgemm_outer($A, $A) 
+    SUITE["matrices"]["ATA_spgemm_outer"][mtx] = @benchmarkable spgemm_outer($A, $A) 
 end
 
 SUITE["indices"] = BenchmarkGroup()
